update to deal with obsoletes better
when looking for a package "foo" to install/update, we have to look
not just for a package named "foo", but also any package that
obsoletes+provides "foo".

So redo the algorithm to satisfy *all* installs/removes from the
property list, so now we don't need to walk the package list at all.